pyproject.toml
[tool.poetry]
name = "documentation"
version = "0.1.0"
description = ""
authors = ["despisemezxc <2003lesha2003@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"
drf-yasg = "^1.21.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
